implemented scons for build and conan for C++ packaging
