+++ library/include/common.hpp
@@ -1,0 +1,23 @@
+#include <Godot.hpp>
+#include "GodotSample.hpp"
+
+
+ /** GDNative Initialize **/
+extern "C" void GDN_EXPORT godot_gdnative_init(godot_gdnative_init_options *o)
+{
+    godot::Godot::gdnative_init(o);
+}
+
+/** GDNative Terminate **/
+extern "C" void GDN_EXPORT godot_gdnative_terminate(godot_gdnative_terminate_options *o)
+{
+    godot::Godot::gdnative_terminate(o);
+}
+
+/** NativeScript Initialize **/
+extern "C" void GDN_EXPORT godot_nativescript_init(void *handle)
+{
+    godot::Godot::nativescript_init(handle);
+
+    godot::register_class<SampleClass>();
+}
